screenshots/Capacity Application/src/api/requirements.txt
azure-cosmos == 4.3.1
azure-identity == 1.12.0
azure-storage-blob == 12.15.0
azure-data-tables == 12.4.2